pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["Igor Ruzhilov <iruzhilov@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.0"
fastapi = "^0.109.0"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
python-dotenv = "^1.0.0"
passlib = "^1.7.4"
pydantic-settings = "^2.1.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
